Move swagger output into src directories
