+++ services/collaboration-service/src/swagger-output.json
